requirements.txt
#web framework
fastapi
uvicorn[standard]
python-multipart # for file upload

# LangChain & EcoSystem
langchain 
langchain-community
langchain-core
langchain-groq
langchain-openai

# Vectorstore
# Crhromadb
pinecone
pinecone-text


openai
cohere
ragas
datasets
loguru

#Embeddings
# sentense-transformers
# langchain-huggingface
langchain-huggingface

# Environment variables
python-dotenv

#PDF parsing
PyPDF # backend for PyPDFLoader via langchain_community

# Typing & Utilities
pydantic
requests
tqdm

# Logging
loguru

# File Uplaod
aiofiles

# Frontend UI
streamlit